add in new sidebands v8 for testing
