+++ make_hists/hms/AntiNuCuts_v8.json
@@ -1,0 +1,119 @@
+{"QElike":[
+    {"name":"NoDeadTime","max":1,"variable":"DeadTime"},
+    {"name":"ApothemX","max":850,"variable":"ApothemX"},
+    {"name":"ApothemY","max":850,"variable":"ApothemY"},
+    {"name":"Z Vertex","min":5980,"max":8422,"variable":"ZVertex"},
+    {"name":"Has MINOS Match","equals":-1,"variable":"IsMinosMatchTrack"}, // -1 for some reason
+    {"name":"NuHelicity","equals":2,"variable":"NuHelicity"},  // 2 for anti-neutrino
+    {"name":"Max Muon Angle in Degrees","max":20.0,"variable":"ThetamuDegrees"},
+    {"name":"Min MuonMomentum","min":1.5,"variable":"PmuGeV"},
+    {"name":"Max MuonMomentum","max":20.0,"variable":"PmuGeV"},
+    {"name":"Max Multiplicity","max":1,"variable":"Multiplicity"},
+    {"name":"GoodRecoil","equals":1,"variable":"GoodRecoil"}
+  ],
+  "QElikeNP":[
+      {"name":"NoDeadTime","max":1,"variable":"DeadTime"},
+      {"name":"ApothemX","max":850,"variable":"ApothemX"},
+      {"name":"ApothemY","max":850,"variable":"ApothemY"},
+      {"name":"Z Vertex","min":5980,"max":8422,"variable":"ZVertex"},
+      {"name":"Has MINOS Match","equals":-1,"variable":"IsMinosMatchTrack"}, // -1 for some reason
+      {"name":"NuHelicity","equals":2,"variable":"NuHelicity"},  // 2 for anti-neutrino
+      {"name":"Max Muon Angle in Degrees","max":20.0,"variable":"ThetamuDegrees"},
+      {"name":"Min MuonMomentum","min":1.5,"variable":"PmuGeV"},
+      {"name":"Max MuonMomentum","max":20.0,"variable":"PmuGeV"},
+      {"name":"Max Multiplicity","max":1,"variable":"Multiplicity"},
+      {"name":"GoodRecoil","equals":1,"variable":"GoodRecoil"}
+    ],
+  "QElikeALLAngles":[
+      {"name":"NoDeadTime","max":1,"variable":"DeadTime"},
+      {"name":"ApothemX","max":850,"variable":"ApothemX"},
+      {"name":"ApothemY","max":850,"variable":"ApothemY"},
+      {"name":"Z Vertex","min":5980,"max":8422,"variable":"ZVertex"},
+      {"name":"Has MINOS Match","equals":-1,"variable":"IsMinosMatchTrack"}, // -1 for some reason
+      {"name":"NuHelicity","equals":2,"variable":"NuHelicity"},  // 2 for anti-neutrino
+      {"name":"Max Muon Angle in Degrees","max":20.0,"variable":"ThetamuDegrees"},
+      {"name":"Min MuonMomentum","min":1.5,"variable":"PmuGeV"},
+      {"name":"Max MuonMomentum","max":20.0,"variable":"PmuGeV"},
+      {"name":"Max Multiplicity","max":1,"variable":"Multiplicity"},
+      {"name":"GoodRecoil","equals":1,"variable":"GoodRecoil"}
+    ],
+  "QElikeFULL":[
+      {"name":"NoDeadTime","max":1,"variable":"DeadTime"},
+      {"name":"ApothemX","max":850,"variable":"ApothemX"},
+      {"name":"ApothemY","max":850,"variable":"ApothemY"},
+      {"name":"Z Vertex","min":5980,"max":8422,"variable":"ZVertex"},
+      {"name":"Has MINOS Match","equals":-1,"variable":"IsMinosMatchTrack"}, // -1 for some reason
+      {"name":"NuHelicity","equals":2,"variable":"NuHelicity"},  // 2 for anti-neutrino
+      {"name":"Max Muon Angle in Degrees","max":20.0,"variable":"ThetamuDegrees"},
+      {"name":"Min MuonMomentum","min":1.5,"variable":"PmuGeV"},
+      {"name":"Max MuonMomentum","max":20.0,"variable":"PmuGeV"},
+      {"name":"Max Multiplicity","max":1,"variable":"Multiplicity"},
+      {"name":"GoodRecoil","equals":1,"variable":"GoodRecoil"}
+    ],
+  "phase_space":[
+    {"name":"TrueApothemX","max":850,"variable":"TrueApothemX"},
+    {"name":"TrueApothemY","max":850,"variable":"TrueApothemY"},
+    {"name":"Z Vertex","min":5980,"max":8422,"variable":"TrueZVertex"},
+    {"name":"TrueMuonMomentumMin","min":1.5,"variable":"TruePmuGeV"},
+    {"name":"TrueMuonMomentumMax","max":20.0,"variable":"TruePmuGeV"},
+    {"name":"Max Muon Angle in Degrees","max":20.0,"variable":"TrueThetamuDegrees"}
+  ],
+  "phase_space_all_angles":[
+    {"name":"TrueApothemX","max":850,"variable":"TrueApothemX"},
+    {"name":"TrueApothemY","max":850,"variable":"TrueApothemY"},
+    {"name":"Z Vertex","min":5980,"max":8422,"variable":"TrueZVertex"},
+    {"name":"TrueMuonMomentumMin","min":1.5,"variable":"TruePmuGeV"},
+    {"name":"TrueMuonMomentumMax","max":20.0,"variable":"TruePmuGeV"}
+  ],
+  "phase_space_full":[
+    {"name":"TrueApothemX","max":850,"variable":"TrueApothemX"},
+    {"name":"TrueApothemY","max":850,"variable":"TrueApothemY"},
+    {"name":"Z Vertex","min":5980,"max":8422,"variable":"TrueZVertex"}
+  ],
+ "qelike":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELike","equals":1,"variable":"IsCCQELike","protonKECutMeV":120.}
+  ],
+ "qelike1p":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELike","equals":1,"variable":"IsCCQELike","protonKECutMeV":120.},
+    {"name":"MCIntType","max":7,"variable":"MCIntType"}
+  ],
+ "qelike2p2h":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELike","equals":1,"variable":"IsCCQELike","protonKECutMeV":120.},
+    {"name":"MCIntType","equals":8,"variable":"MCIntType"}
+  ],
+  "qelikenot":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELike","equals":0,"variable":"IsCCQELike","protonKECutMeV":120.}
+  ],
+  "qelikenp":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELikeAll","equals":1,"variable":"IsCCQELikeAll"}
+  ],
+  "qelike1p_np":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELikeAll","equals":1,"variable":"IsCCQELikeAll"},
+    {"name":"MCIntType","max":7,"variable":"MCIntType"}
+  ],
+  "qelike2p2h_np":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELikeAll","equals":1,"variable":"IsCCQELikeAll"},
+    {"name":"MCIntType","equals":8,"variable":"MCIntType"}
+  ],
+  "qelikenot_np":[
+    {"name":"PDG","equals":-14,"variable":"TruthNuPDG"},
+    {"name":"IsCC","equals":1,"variable":"IsCC"},
+    {"name":"IsCCQELikeAll","equals":0,"variable":"IsCCQELikeAll"}
+  ],
+  "data":[] // data doesn't cut anything at truth level
+}
+
