+++ make_hists/hms/CCQEMAT_LinkDef.h
@@ -1,0 +1,9 @@
+#ifdef __CINT__
+
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+
+#endif
